Add more notes on required installation / setup, fix cargo config, fix vga buffer addr

+++ src/main.rs
@@ -9,7 +9,7 @@
 
 use core::panic::PanicInfo;
 
-const VGA_BUFFER_ADDRESS: u8 = 0xb8000;
+const VGA_BUFFER_ADDRESS: u32 = 0xb8000;
 const OS_WELCOME_MSG: &[u8] = b"Hello World!";
 
 #[panic_handler]
